debuggin randomize_target_small_task

+++ g6k/decl.pxd
@@ -292,6 +292,6 @@ cdef extern from "../kernel/slicer.h" nogil:
     cdef cppclass RandomizedSlicer:
         RandomizedSlicer(Siever &sieve, unsigned long int seed)
         #void grow_db_with_target( std::array<LFT,128> &t_yr, size_t n_per_target);
-        void grow_db_with_target( float* t_yr, size_t n_per_target);
+        void grow_db_with_target( double* t_yr, size_t n_per_target);
